add support for pkg build dry-run
